@@ -1,4 +1,4 @@
-    const mongoose =require("mongoose");
+    import mongoose from 'mongoose';
 
     const CartSchema = new mongoose.Schema(
         {
@@ -18,4 +18,4 @@
         {timestamps: true}
     );
 
-    module.exports=mongoose.model("Cart",CartSchema);
+    export default mongoose.model('Cart', CartSchema);
