@@ -1,4 +1,4 @@
-const mongoose = require("mongoose");
+import mongoose from 'mongoose';
 
 const ProductSchema = new mongoose.Schema(
     {
@@ -28,4 +28,4 @@ const ProductSchema = new mongoose.Schema(
     { timestamps: true }
 );
 
-module.exports = mongoose.model("Product", ProductSchema);
+export default mongoose.model('Product', ProductSchema);
