@@ -1,17 +1,36 @@
 // routes/razorpay.js
-const express = require("express");
-const Razorpay = require("razorpay");
-const crypto = require("crypto");
+import express from 'express';
+import Razorpay from 'razorpay';
+import crypto from 'crypto';
+import { paymentRateLimit } from '../middleware/arcjet.js';
 
 const router = express.Router();
 
-const razorpay = new Razorpay({
-  key_id: process.env.RAZORPAY_KEY_ID,
-  key_secret: process.env.RAZORPAY_SECRET,
-});
+let razorpay = null;
+
+const ensureRazorpay = () => {
+  const { RAZORPAY_KEY_ID, RAZORPAY_SECRET } = process.env;
+
+  if (!RAZORPAY_KEY_ID || !RAZORPAY_SECRET) {
+    console.warn(
+      "Razorpay credentials are missing. Set RAZORPAY_KEY_ID and RAZORPAY_SECRET in your environment."
+    );
+    razorpay = null;
+    return null;
+  }
+
+  if (!razorpay) {
+    razorpay = new Razorpay({
+      key_id: RAZORPAY_KEY_ID,
+      key_secret: RAZORPAY_SECRET,
+    });
+  }
+
+  return razorpay;
+};
 
 // Create Razorpay Order
-router.post("/order", async (req, res) => {
+router.post("/order", paymentRateLimit, async (req, res) => {
   try {
     const { amount, currency = "INR" } = req.body;
 
@@ -21,7 +40,15 @@ router.post("/order", async (req, res) => {
       receipt: `receipt_order_${Math.random() * 10000}`,
     };
 
-    const order = await razorpay.orders.create(options);
+    const client = ensureRazorpay();
+
+    if (!client) {
+      return res
+        .status(500)
+        .json({ message: "Razorpay is not configured on the server." });
+    }
+
+    const order = await client.orders.create(options);
 
     if (!order) return res.status(500).send("Order creation failed");
     res.json(order);
@@ -32,8 +59,17 @@ router.post("/order", async (req, res) => {
 });
 
 // Validate Payment Signature
-router.post("/order/validate", async (req, res) => {
-  const { razorpay_order_id, razorpay_payment_id, razorpay_signature } = req.body;
+router.post("/order/validate", paymentRateLimit, async (req, res) => {
+  const client = ensureRazorpay();
+
+  if (!client) {
+    return res
+      .status(500)
+      .json({ message: "Razorpay is not configured on the server." });
+  }
+
+  const { razorpay_order_id, razorpay_payment_id, razorpay_signature } =
+    req.body;
 
   const sha = crypto.createHmac("sha256", process.env.RAZORPAY_SECRET);
   sha.update(`${razorpay_order_id}|${razorpay_payment_id}`);
@@ -50,4 +86,4 @@ router.post("/order/validate", async (req, res) => {
   });
 });
 
-module.exports = router;
+export default router;
