@@ -1,5 +1,5 @@
 // verifyToken.js - Enhanced for farmer-retailer system
-const jwt = require("jsonwebtoken");
+import jwt from 'jsonwebtoken';
 
 const verifyToken = (req, res, next) => {
     const authHeader = req.headers.token;
@@ -84,7 +84,7 @@ const verifyFarmerProduct = (req, res, next) => {
     });
 };
 
-module.exports = {
+export {
     verifyToken,
     verifyTokenAndAuthorization,
     verifyTokenAndAdmin,
